requirements.txt
flask
line-bot-sdk
google-generativeai
aiohttp
requests
python-dotenv
gunicorn
Werkzeug
psutil
pytz
schedule